Folder structure refactored
