+++ .gitignore
@@ -107,5 +107,4 @@ ENV/
 # self-generating
 Project/Server/DATA
 *.db
-
-dump\.rdb
+*.rdb
